fixed some stuff

+++ src/app/actions/moodActions.tsx
@@ -6,10 +6,12 @@ import { getServerSession } from 'next-auth';
 
 export const getUserMoods = async (): Promise<Mood[]> => {
   const session = await getServerSession(authOptions);
+
   if (!session?.user?.id) {
     console.error('User is not authenticated or missing user ID');
-    return [];
+    throw new Error('User is not authenticated');
   }
+
   const userId = session.user.id;
 
   return await db.mood.findMany({
@@ -19,36 +21,38 @@ export const getUserMoods = async (): Promise<Mood[]> => {
 
 export const addUserMood = async (mood: Mood): Promise<Mood> => {
   const session = await getServerSession(authOptions);
+
   if (!session?.user?.id) {
     console.error('User is not authenticated or missing user ID');
-    return mood;
+    throw new Error('User is not authenticated');
   }
 
   const userId = session.user.id;
   mood.userId = userId;
   delete (mood as { id?: string }).id;
+
   return await db.mood.create({ data: mood });
 };
 
 export const deleteUserMood = async (mood: Mood): Promise<Mood | null> => {
   const session = await getServerSession(authOptions);
+
   if (!session?.user?.id) {
     console.error('User is not authenticated or missing user ID');
-    return null;
+    throw new Error('User is not authenticated');
   }
 
   const userId = session.user.id;
   mood.userId = userId;
 
   try {
-    // Delete the mood and return the deleted mood object
     const deletedMood = await db.mood.delete({
       where: {
-        id: mood.id,  // Make sure the mood has an id
+        id: mood.id,
+        userId: userId, // is this line redundant here?
       },
     });
-
-    return deletedMood; // Return the deleted mood object
+    return deletedMood;
   } catch (error) {
     console.error('Error deleting mood:', error);
     return null;
@@ -56,11 +60,19 @@ export const deleteUserMood = async (mood: Mood): Promise<Mood | null> => {
 };
 
 export const deleteUserMoodsBulk = async (moodIds: string[]) => {
+  const session = await getServerSession(authOptions);
+
+  if (!session?.user?.id) {
+    console.error('User is not authenticated or missing user ID');
+    throw new Error('User is not authenticated');
+  }
+  const userId = session.user.id;
+
   try {
-    // Using Prisma's deleteMany method
     await db.mood.deleteMany({
       where: {
-        id: { in: moodIds }, // Use 'in' without the '$'
+        id: { in: moodIds },
+        userId: userId,
       },
     });
     return true;
